requirements.txt
streamlit==1.45.0
torch==2.7.1+cu128
torchvision==0.22.1+cu128
pillow==11.1.0
